\name{beta.os_prime}
\alias{beta.os_prime}
\title{Measure the distance between a network and its metaweb}
\usage{
beta.os_prime(W, ...)
}
\arguments{
  \item{W}{a list of networks}

  \item{...}{additional arguments to be passed to
  \link{betalink}}
}
\value{
An array of the values of Beta OS'
}
\description{
Returns the values of beta OS', i.e. the distace between
all realizations, and the revelant subset from the metaweb
}

