% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_os_prime.R
\name{beta_os_prime}
\alias{beta_os_prime}
\title{Measure the distance between a network and its metaweb}
\usage{
beta_os_prime(N, ...)
}
\arguments{
\item{N}{a list of networks}

\item{...}{additional arguments to be passed to \link{betalink}}
}
\value{
An array of the values of Beta OS'
}
\description{
Returns the values of beta OS', i.e. the distace between all realizations, and the revelant subset from the metaweb
}

