% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/betalink.r
\docType{data}
\name{clownfishes}
\alias{clownfishes}
\title{Anemone/fish interaction data}
\format{16 adjancency matrices with species names}
\description{
Anemone/fish interaction data
}
\keyword{dataset}

