% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_networks.R
\name{df_from_A}
\alias{df_from_A}
\title{data.frame from adjancency matrix}
\usage{
df_from_A(A)
}
\arguments{
\item{A}{an adjacency matrix}
}
\description{
Transforms an Adjacency matrix into a data frame
}

