% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metaweb.r
\name{metaweb}
\alias{metaweb}
\title{Returns a metaweb given a list of networks}
\usage{
metaweb(n)
}
\arguments{
\item{n}{a \code{list} of graphs}
}
\description{
Given a list of networks, this function returns the metaweb
}

