% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betalink.r
\name{name_networks}
\alias{name_networks}
\title{Give names to networks}
\usage{
name_networks(w)
}
\arguments{
\item{w}{A list (of networks, but who am I to judge?)}
}
\description{
If the networks (in a list) have no names, give them names
}

