% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_betaplot.R
\name{network_betaplot}
\alias{network_betaplot}
\title{Plot a network with species and interactions highlighted}
\usage{
network_betaplot(n1, n2, na = "#2ca02c", nb = "#1f77b4", ns = "grey", ...)
}
\arguments{
\item{n1}{a network}

\item{n2}{a second network}

\item{na}{color of items unique to network 1}

\item{nb}{color of items unique to network 2}

\item{ns}{color of shared items}

\item{...}{additional arguments to be passed to plot}
}
\value{
Nothing
}
\description{
Plot differences between two networks
}

