\name{null.metaweb} 
\Rdversion{1.1} 
\alias{null.metaweb} 
\title{Null model based on the metaweb}
\description{Generates a series of null networks based on the metaweb.}
\usage{ 
null.metaweb(template,reps)
}

\arguments{ 
\item{template}{A probability matrix}
\item{reps}{Number of networks to generate} 
}

\value{
	\item{}{A list of \kbd{reps} networks with the same rows and colums names than the template -- species keep identical names}
}
 
\details{---}

\references{---}

\examples{} 