% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betalink.r
\docType{data}
\name{anemonefish}
\alias{anemonefish}
\title{Anemone/fish interaction data}
\format{16 adjacency matrices with species names}
\description{
Anemone/fish interaction data
}
\keyword{dataset}

