\name{network_betadiversity}
\alias{network_betadiversity}
\title{Components of beta-diversity for a list of networks}
\usage{
network_betadiversity(N, ...)
}
\arguments{
  \item{N}{a list of networks}

  \item{...}{additional arguments to be passed to
  \link{betalink}}
}
\value{
A dataframe with the pairwise distances
}
\description{
Given a list of networks, returns the pairwise
beta-diversity components
}

