% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_betadiversity.R
\name{network_betadiversity}
\alias{network_betadiversity}
\title{Components of beta-diversity for a list of networks}
\usage{
network_betadiversity(N, complete = FALSE, ...)
}
\arguments{
\item{N}{a list of networks}

\item{complete}{(boolean) whether all combinations of networks should be tested}

\item{...}{additional arguments to be passed to \link{betalink}}
}
\value{
A dataframe with the pairwise distances
}
\description{
Given a list of networks, returns the pairwise beta-diversity components
}

