% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_networks.R
\name{prepare_networks}
\alias{prepare_networks}
\title{Prepare networks}
\usage{
prepare_networks(w, directed = TRUE)
}
\arguments{
\item{w}{A list of network matrices}

\item{directed}{whether the edges are directed or not}
}
\description{
Taking a list of networks as matrices, returns a list of igraph objects
}
\examples{
data(anemonefish)
networks <- prepare_networks(anemonefish, TRUE)
print(networks$Timur)
}

